requirements.txt
pymovie
praw
requests
pillow
gtts
pydub
pyrubberband